Project Mega Integration: Framework
Integrate the wood from many, many mods. First, we need shells for this.

Included a dummy file in one of the folders, just so this can be
committed.

YOU GET A BEE BOX! AND YOU GET A BEE BOX! AND YOU GET A BEE BOX!
EVERYONE GETS A BEE BOX!