BY ATOMIC FIRE
I SHALL PURGE MY ENEMIES AND GAIN THEIR CAPACITIES FOR MANIPULATION,
NARCISSISM, AND SUCH, TO BE USED AGAINST FUTURE ENEMIES

FUCK YOU ALL HAHAHAHAHAHAHAHAHAHAHAHAHAHAHAHAHA

And I will execute great vengeance upon them with furious rebukes; and
they shall know that I am the Lord, when I shall lay my vengeance upon
them.

Then the third angel sounded his trumpet, and a great star burning like
a torch fell from heaven and landed on a third of the rivers and on the
springs of water.

The name of the star is Wormwood. A third of the waters turned bitter
like wormwood oil, and many people died from the bitter waters.

Then the fourth angel sounded his trumpet, and a third of the sun and
moon and stars were struck. A third of the stars were darkened, a third
of the day was without light, and a third of the night as well.

what am i doing